Complete user collection page and modify endpoint

+++ ramen-pocket/Models/Store.swift
@@ -29,7 +29,7 @@ struct Store: Codable, Identifiable {
 struct BusinessHour: Codable {
     let off: Bool
     let begin: Double
-    let end: Int
+    let end: Double
 }
 
 // MARK: - Course
